requirements.txt
ultralytics==8.0.193
Pillow==10.0.1
easyocr==1.7.1
paddleocr==2.7.0.3